Local change

+++ src/main/java/com/task/api/dao/ParentTaskRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.task.api.entity.ParentTask;
 
 @Repository
-public interface ParentTaskRepository extends CrudRepository<ParentTask, Integer> {
+public interface ParentTaskRepository extends CrudRepository<ParentTask, Long> {
 
 }
