+++ src/main/java/com/task/api/dao/ProjectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.task.api.entity.Project;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Integer> {
+public interface ProjectRepository extends CrudRepository<Project, Long> {
 
 }
